requirements.txt
python-pptx==0.6.23
openai==1.12.0
python-dotenv==1.0.1
httpx==0.27.0 